Extract RepairEvent into separate class
